+++ src/Bican/Roles/RolesServiceProvider.php
@@ -12,10 +12,9 @@ class RolesServiceProvider extends ServiceProvider {
 	public function boot()
 	{
 		$this->publishes([
-			__DIR__.'/../../migrations/' => base_path('/database/migrations')
-		], 'migrations');
-        
-        $this->publishes([__DIR__ . '/../../config/roles.php' => config_path('roles.php')]);
+			__DIR__ . '/../../migrations/' => base_path('/database/migrations'),
+            __DIR__ . '/../../config/roles.php' => config_path('roles.php')
+		]);
 
         $this->registerBladeExtensions();
     }
