+++ src/config/roles.php
@@ -1,15 +1,54 @@
 <?php
-    /*Config file for Bican/Roles package*/
 
 return [
-    /*pretend lets you simulate how roles package behave*/
 
-    /*enabled (true|false) tells package to simulate behaviour according to options array*/
-    'pretend' => ['enabled' => 'false',
-                  'options' => ['hasPermission' => 'true', //hasPermission always returns assigned value (true|false)
-                                'hasRole'       => 'true', //hasRole always returns assigned value (true|false)
-                                'allowed'       => 'true', //allowed always returns assigned value (true|false)
-                                'is'            => 'true'] //is always returns assigned value (true|false)
-    ]
+    /*
+	|--------------------------------------------------------------------------
+	| Package Connection
+	|--------------------------------------------------------------------------
+	|
+	| You can set a different database connection for this package. It will set
+	| new connection for models Role and Permission. When this option is null,
+	| it will connect to the main database, which is set up in database.php
+	|
+	*/
+
+    'connection' => null,
+
+    /*
+	|--------------------------------------------------------------------------
+	| Slug Separator
+	|--------------------------------------------------------------------------
+	|
+	| Here you can change the slug separator. This is very important in matter
+	| of magic method __call() and also a SlugableTrait. The default value
+	| is a dot.
+	|
+	*/
+
+    'separator' => '.',
+
+    /*
+	|--------------------------------------------------------------------------
+	| Roles, Permissions and Allowed "Pretend"
+	|--------------------------------------------------------------------------
+	|
+	| You can pretend or simulate package behavior no matter what is in your
+	| database. It is really useful when you are testing you application.
+	| Set up what will methods is(), can() and allowed() return.
+	|
+	*/
+
+    'pretend' => [
+
+        'enabled' => false,
+
+        'options' => [
+            'is'        => true,
+            'can'       => true,
+            'allowed'   => true,
+        ],
+
+    ],
+
 ];
-
